add new argument-less configuration syntax
The old way of configuring bunyan was to pass an
argument to the block passed to the .configure method.
This seemed a little too verbose, considering that
all attributes set in the config block are set on
the config object anyway. 

I think this keeps the configuration DSL a bit cleaner.